+++ next-docs/CHANGELOG.md
@@ -1,5 +1,57 @@
 # Change Log
 
+## 9.21.2
+
+### Patch Changes
+
+- Bump patch version
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.21.2
+  - @heathmont/moon-accordion@9.21.2
+  - @heathmont/moon-assets@9.21.2
+  - @heathmont/moon-charts@9.21.2
+  - @heathmont/moon-components@9.21.2
+  - @heathmont/moon-core@9.21.2
+  - @heathmont/moon-datepicker@9.21.2
+  - @heathmont/moon-draggabletable@9.21.2
+  - @heathmont/moon-icons@9.21.2
+  - @heathmont/moon-pagination@9.21.2
+  - @heathmont/moon-popover@9.21.2
+  - @heathmont/moon-select@9.21.2
+  - @heathmont/moon-sidebar@9.21.2
+  - @heathmont/moon-table@9.21.2
+  - @heathmont/moon-themes@9.21.2
+  - @heathmont/moon-utils@9.21.2
+  - @heathmont/moon-icons-tw@9.21.2
+  - @heathmont/moon-table-tw@9.21.2
+  - @heathmont/moon-themes-tw@9.21.2
+
+## 9.21.1
+
+### Patch Changes
+
+- Bump patch
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.21.1
+  - @heathmont/moon-accordion@9.21.1
+  - @heathmont/moon-assets@9.21.1
+  - @heathmont/moon-charts@9.21.1
+  - @heathmont/moon-components@9.21.1
+  - @heathmont/moon-core@9.21.1
+  - @heathmont/moon-datepicker@9.21.1
+  - @heathmont/moon-draggabletable@9.21.1
+  - @heathmont/moon-icons@9.21.1
+  - @heathmont/moon-pagination@9.21.1
+  - @heathmont/moon-popover@9.21.1
+  - @heathmont/moon-select@9.21.1
+  - @heathmont/moon-sidebar@9.21.1
+  - @heathmont/moon-table@9.21.1
+  - @heathmont/moon-themes@9.21.1
+  - @heathmont/moon-utils@9.21.1
+  - @heathmont/moon-icons-tw@9.21.1
+  - @heathmont/moon-table-tw@9.21.1
+  - @heathmont/moon-themes-tw@9.21.1
+
 ## 9.21.0
 
 ### Minor Changes