+++ workspaces/tables/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @heathmont/moon-table-tw
 
+## 9.21.2
+
+### Patch Changes
+
+- Bump patch version
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.21.2
+
+## 9.21.1
+
+### Patch Changes
+
+- Bump patch
+- Updated dependencies [undefined]
+  - @heathmont/moon-core-tw@9.21.1
+
 ## 9.21.0
 
 ### Minor Changes