+++ workspaces/templates/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @heathmont/moon-templates
 
+## 9.21.2
+
+### Patch Changes
+
+- Bump patch version
+
+## 9.21.1
+
+### Patch Changes
+
+- Bump patch
+
 ## 9.21.0
 
 ### Minor Changes